refactor(rust): Reorganize `polars_io::parquet` module (#15860)
