+++ crates/polars-io/src/parquet/write/mod.rs
@@ -1,0 +1,10 @@
+//! Functionality for reading and writing Apache Parquet files.
+
+mod batched_writer;
+mod options;
+mod writer;
+
+pub use batched_writer::BatchedWriter;
+pub use options::{BrotliLevel, GzipLevel, ParquetCompression, ParquetWriteOptions, ZstdLevel};
+pub use polars_parquet::write::RowGroupIter;
+pub use writer::ParquetWriter;
