+++ crates/polars-lazy/src/prelude.rs
@@ -5,7 +5,7 @@ pub use polars_io::ipc::IpcWriterOptions;
 #[cfg(feature = "json")]
 pub use polars_io::json::JsonWriterOptions;
 #[cfg(feature = "parquet")]
-pub use polars_io::parquet::ParquetWriteOptions;
+pub use polars_io::parquet::write::ParquetWriteOptions;
 pub use polars_ops::prelude::{JoinArgs, JoinType, JoinValidation};
 #[cfg(feature = "rank")]
 pub use polars_ops::prelude::{RankMethod, RankOptions};
