+++ crates/polars-pipe/src/executors/sources/parquet.rs
@@ -9,7 +9,8 @@ use polars_core::error::*;
 use polars_core::prelude::Series;
 use polars_core::POOL;
 use polars_io::cloud::CloudOptions;
-use polars_io::parquet::{BatchedParquetReader, FileMetaData, ParquetOptions, ParquetReader};
+use polars_io::parquet::metadata::FileMetaDataRef;
+use polars_io::parquet::read::{BatchedParquetReader, ParquetOptions, ParquetReader};
 use polars_io::pl_async::get_runtime;
 use polars_io::predicates::PhysicalIoExpr;
 use polars_io::prelude::materialize_projection;
@@ -36,7 +37,7 @@ pub struct ParquetSource {
     file_options: FileScanOptions,
     #[allow(dead_code)]
     cloud_options: Option<CloudOptions>,
-    metadata: Option<Arc<FileMetaData>>,
+    metadata: Option<FileMetaDataRef>,
     file_info: FileInfo,
     verbose: bool,
     run_async: bool,
@@ -190,7 +191,7 @@ impl ParquetSource {
         paths: Arc<[PathBuf]>,
         options: ParquetOptions,
         cloud_options: Option<CloudOptions>,
-        metadata: Option<Arc<FileMetaData>>,
+        metadata: Option<FileMetaDataRef>,
         file_options: FileScanOptions,
         file_info: FileInfo,
         verbose: bool,
