+++ crates/polars-plan/src/logical_plan/functions/count.rs
@@ -7,9 +7,9 @@ use polars_io::cloud::CloudOptions;
 #[cfg(feature = "csv")]
 use polars_io::csv::read::count_rows as count_rows_csv;
 #[cfg(all(feature = "parquet", feature = "cloud"))]
-use polars_io::parquet::ParquetAsyncReader;
+use polars_io::parquet::read::ParquetAsyncReader;
 #[cfg(feature = "parquet")]
-use polars_io::parquet::ParquetReader;
+use polars_io::parquet::read::ParquetReader;
 #[cfg(all(feature = "parquet", feature = "async"))]
 use polars_io::pl_async::{get_runtime, with_concurrency_budget};
 #[cfg(feature = "parquet")]
